Roll Open Screen from 448a19d1f24e to c8117813dfbf (3 revisions)
https://chromium.googlesource.com/openscreen.git/+log/448a19d1f24e..c8117813dfbf

2026-04-01 jophba@chromium.org [Cast Streaming] Support kNumFramesDroppedByEncoder
2026-03-31 jophba@chromium.org [Code Health] Cleanup Sender, Receiver ownership
2026-03-31 jophba@chromium.org [OWNERS] Add Taesun Yeoum as an additional reviewer

This patch includes some fixes for breaking Open Screen changes.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7718344
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1611261}

+++ components/cast_streaming/browser/frame/stream_consumer.h
@@ -99,7 +99,7 @@ class StreamConsumer final : public openscreen::cast::Receiver::Consumer {
   void MaybeSendNextFrame();
 
   // openscreen::cast::Receiver::Consumer implementation.
-  void OnFramesReady(int next_frame_buffer_size) override;
+  void OnFramesReady(size_t next_frame_buffer_size) override;
 
   // This receiver should skip all frames with id less than this value. Set by a
   // call to FlushUntil() and 0 when no flush is ongoing.
